requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.6.1
async-timeout==4.0.2
bcrypt==3.2.2
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
Deprecated==1.2.13
deta==1.1.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fakeredis==1.8.1
fastapi==0.78.0
fastapi-mail==1.0.9
gunicorn==20.0.4
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
peewee==3.15.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==4.3.4
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.19.1
starlette-context==0.3.4
typing_extensions==4.2.0
uvicorn==0.18.1
watchfiles==0.15.0
websockets==10.3
wrapt==1.14.1
